move getShape to options
